+++ spec/shibaraku_spec.rb
@@ -1,0 +1,7 @@
+require "spec_helper"
+
+describe Shibaraku do
+  it "has a version number" do
+    expect(Shibaraku::VERSION).not_to be nil
+  end
+end